Standard compliance for mutex types
- Provides fully standard-compliant `mutex` class if targeting Windows 7
or higher. Has constexpr constructor and is a standard layout type.
- Provides fully standard-compliant `mutex` class for Windows XP by
using deferred initialization.
- Makes `timed_mutex` standard-compliant for all Windows targets, but at
the expense of a slight memory footprint increase for
`recursive_timed_mutex`.
- Alters non-recursive error checking to achieve exception-safety
specification for `mutex` and `timed_mutex`. Specifically, the `lock()`
function should not acquire the lock if it happens to throw an
exception. Note that this is technically unnecessary, as it is covered
under "undefined behavior", but it makes non-terminating behavior more
consistent.
- With `mutex` now standard compliant, once_flag no longer needs the
atomic implementation, so it has been removed.