Add -fno-exceptions support (#53)
* Add -fno-exceptions support

* Change <c++/bits/exception_defines.h> include to <bits/exception_defines.h>

* Change:
__throw_exception_again
by
__throw_<<exception_name_here>>

* Change to a more flexible "throw" system

* More robust detection of -fno-exceptions.

If the __cpp_exceptions macro is defined, use it. Otherwise, check for
compiler-specific feature macros (__EXCEPTIONS for GCC) in case an old
version of GCC is being used. If we are unsure, we must assume that C++
exceptions are being used, as this is the standard behavior.

* Create an exception pointer without catch/throw.

Use an appropriate standard-library function to create exception
pointers without using catch or throw clauses.

* Fix many compilation errors.

Despite the intention of previous authors' contributions, the
contributed code did not, in fact, compile with -fno-exceptions set.

This updates all headers so that they will successfully compile
regardless of whether -fno-exceptions is specified (tested in MinGW
GCC).