Reduce reliance on compiler intelligence.
Previously, a copy of an integer sequence was intantiated and passed as
an argument to ThreadFuncCall::callFunc. This provided an integer
sequence 0 .. N using a function template. This commit alters the
behavior of ThreadFuncCall and its point of use so that the correct
function is called directly.

This commit should remove the small chance that a compiler might miss the
optimization provided by inlining the innermost function.