Miscellaneous improvements (#61)
* Silence a warning.

Note: Style warning only. No change to minor or major version.

* Reduce reliance on compiler intelligence.

Previously, a copy of an integer sequence was intantiated and passed as
an argument to ThreadFuncCall::callFunc. This provided an integer
sequence 0 .. N using a function template. This commit alters the
behavior of ThreadFuncCall and its point of use so that the correct
function is called directly.

This commit should remove the small chance that a compiler might miss the
optimization provided by inlining the innermost function.

* Improve type safety in implementation

Pass a pointer to the correct type (and convert later) to avoid strict aliasing violations.

* Remove last non-standard extension from thread::id

According to the standard, a thread::id is only able to be constructed with a non-thread value, unless requested by thread::get_id or this_thread::get_id. This commit removes all non-standard public access to the thread::id's value.

* Add regression test for a property of thread::id

thread::id is required to be TriviallyCopyable. The tests in tests.cpp now include a compile-time test for this property.