Prevent `async` from blocking `future::wait_for` (#77)
* Fix locking in async

Calling wait, wait_for, etc. in future or shared_future could lead to
waiting for an asynchronous function to complete, as it would hold the
mutex during its computation. Fix this by invoking asynchronous
functions prior to locking the mutex, storing their results only when
done.

* Add test for timeouts

Prevent regressions in `async` by adding a test that requires a minimum
of 4 timeouts (out of expected 9).